Atualizalçao atividade

+++ NumeroInformado.java
@@ -3,7 +3,7 @@ public class NumeroInformado {
     public static void main(String[] args){
         Scanner scanner = new Scanner(System.in);
          System.out.println("Digite um número: ");
-         int N1 = scanner.nextInt();
+         double N1 = scanner.nextDouble();
          System.out.println("O número digitado foi: " + N1);
         scanner.close();
     }
